+++ libs/fakeblocks/src/fakeblocks/ui/vtk/__init__.py
@@ -1,3 +1,0 @@
-from capytaine.ui.vtk.mesh_viewer import MeshViewer
-from capytaine.ui.vtk.body_viewer import FloatingBodyViewer
-from capytaine.ui.vtk.animation import Animation