+++ src/capytaine/bodies/bodies.py
@@ -73,10 +73,8 @@ class FloatingBody(_HydrostaticsMixin, AbstractBody):
     def __init__(self, mesh=None, dofs=None, *, lid_mesh=None, center_of_mass=None, mass=None, name=None):
         if mesh is None:
             self.mesh = Mesh(name="dummy_mesh")
-        elif isinstance(mesh, AbstractMesh):
-            self.mesh = mesh
         else:
-            raise TypeError("Unrecognized `mesh` object passed to the FloatingBody constructor.")
+            self.mesh = mesh
 
         if lid_mesh is None:
             self.lid_mesh = None
@@ -145,9 +143,10 @@ class FloatingBody(_HydrostaticsMixin, AbstractBody):
         mesh = load_mesh(filename, file_format, name=f"{name}_mesh")
         return FloatingBody(mesh, name=name)
 
-    def __lt__(self, other: 'FloatingBody') -> bool:
-        """Arbitrary order. The point is to sort together the problems involving the same body."""
-        return self.name < other.name
+    @property
+    def bodies(self):
+        """For consistency with Multibody"""
+        return [self]
 
     ##########
     #  Dofs  #
@@ -281,6 +280,9 @@ class FloatingBody(_HydrostaticsMixin, AbstractBody):
             LOG.debug(f"Copy {self.name} under the name {name}.")
         return new_body
 
+    def rename(self, name: 'str') -> 'FloatingBody':
+        return self.copy(name=name)
+
     def assemble_regular_array(self, distance, nb_bodies):
         """Create an regular array of identical bodies.
 
@@ -296,8 +298,7 @@ class FloatingBody(_HydrostaticsMixin, AbstractBody):
         FloatingBody
         """
         bodies = (self.translated((i*distance, j*distance, 0), name=f"{i}_{j}") for j in range(nb_bodies[1]) for i in range(nb_bodies[0]))
-        array = FloatingBody.join_bodies(*bodies)
-        array.name = f"array_of_{self.name}"
+        array = FloatingBody.join_bodies(*bodies, name=f"array_of_{self.name}")
         return array
 
     def assemble_arbitrary_array(self, locations:np.ndarray):
@@ -308,7 +309,7 @@ class FloatingBody(_HydrostaticsMixin, AbstractBody):
         fb_list = []
         for idx, li in enumerate(locations):
             fb_list.append(self.translated(np.append(li, 0), name='arbitrary_array_body{:02d}'.format(idx)))
-        arbitrary_array = FloatingBody.join_bodies(*fb_list)
+        arbitrary_array = FloatingBody.join_bodies(*fb_list, name=f"array_of_{self.name}")
 
         return arbitrary_array
 
@@ -436,6 +437,7 @@ class FloatingBody(_HydrostaticsMixin, AbstractBody):
             name=name
         )
 
+    @lru_cache
     def immersed_part(self, free_surface=0.0, *, sea_bottom=None, water_depth=None, name=None) -> "FloatingBody":
         clipped_mesh, clipped_lid_mesh, updated_dofs = self._apply_on_mesh(
             self.mesh.__class__.immersed_part,
@@ -510,28 +512,7 @@ class FloatingBody(_HydrostaticsMixin, AbstractBody):
         yield "name", self.name
 
     def animate(self, motion, *args, **kwargs):
-        """Display a motion as a 3D animation.
-
-        Parameters
-        ==========
-        motion: dict or pd.Series or str
-            A dict or series mapping the name of the dofs to its amplitude.
-            If a single string is passed, it is assumed to be the name of a dof
-            and this dof with a unit amplitude will be displayed.
-        """
-        from capytaine.ui.vtk.animation import Animation
-        if isinstance(motion, str):
-            motion = {motion: 1.0}
-        elif isinstance(motion, xr.DataArray):
-            motion = {k: motion.sel(radiating_dof=k).data for k in motion.coords["radiating_dof"].data}
-
-        if any(dof not in self.dofs for dof in motion):
-            missing_dofs = set(motion.keys()) - set(self.dofs.keys())
-            raise ValueError(f"Trying to animate the body {self.name} using dof(s) {missing_dofs}, but no dof of this name is defined for {self.name}.")
-
-        animation = Animation(*args, **kwargs)
-        animation._add_actor(self.mesh.merged(), faces_motion=sum(motion[dof_name] * dof for dof_name, dof in self.dofs.items() if dof_name in motion))
-        return animation
+        raise NotImplementedError("Not yet re-implemented in version 3")
 
     #################################
     # Irregular frequencies removal #