+++ src/capytaine/bodies/dofs.py
@@ -8,24 +8,41 @@ import logging
 import numpy as np
 import xarray as xr
 
+from capytaine.meshes.abstract_meshes import AbstractMesh
+
 LOG = logging.getLogger(__name__)
 
 
 class AbstractDof(ABC):
     @lru_cache
-    def evaluate_motion(self, mesh):
+    def evaluate_motion(self, mesh: AbstractMesh) -> np.ndarray:
         if mesh.nb_faces == 0:
             return np.empty((mesh.nb_faces, 3))
         else:
             return self.evaluate_motion_at_points(mesh.faces_centers)
 
     @abstractmethod
-    def evaluate_motion_at_points(self, points):
+    def evaluate_motion_at_points(self, points: np.ndarray) -> np.ndarray:
         # points is an array of shape (nb_points, 3)
+        # output is of shape (nb_points, 3)
         ...
 
+    @lru_cache
+    def evaluate_gradient_of_motion(self, mesh: AbstractMesh) -> np.ndarray:
+        # output is of shape (nb_points, 3, 3)
+        if mesh.nb_faces == 0:
+            return np.empty((mesh.nb_faces, 3, 3))
+        else:
+            return self.evaluate_gradient_of_motion_at_points(mesh.faces_centers)
+
     @abstractmethod
-    def evaluate_gradient_of_motion(self, mesh):
+    def evaluate_gradient_of_motion_at_points(self, points: np.ndarray) -> np.ndarray:
+        # points is an array of shape (nb_points, 3)
+        # output is of shape (nb_points, 3, 3)
+        # output is a Jacobian matrix, such that output[i_point, i_dir, i_deriv_dir]
+        # is the derivative with respect to `i_deriv_dir` of the `i_dir` component of the motion.
+        # In other words, output[:, 0, :] is the gradient of the x-component of the motion on each face and
+        # output[:, :, 0] is the derivative with respect to x of the motion vector on each face.
         ...
 
 
@@ -37,12 +54,11 @@ class TranslationDof(AbstractDof):
     def __str__(self):
         return f"TranslationDof(direction={self.direction})"
 
-    def evaluate_motion_at_points(self, points) -> np.array:
+    def evaluate_motion_at_points(self, points: np.ndarray) -> np.ndarray:
         return np.tile(self.direction, (points.shape[0], 1))
 
-    @lru_cache
-    def evaluate_gradient_of_motion(self, mesh) -> np.array:
-        return np.zeros((mesh.nb_faces, 3, 3))
+    def evaluate_gradient_of_motion_at_points(self, points: np.ndarray) -> np.ndarray:
+        return np.zeros((points.shape[0], 3, 3))
 
 
 class RotationDof(AbstractDof):
@@ -60,31 +76,44 @@ class RotationDof(AbstractDof):
     def __str__(self):
         return f"RotationDof(rotation_center={self.rotation_center}, direction={self.direction})"
 
-    def evaluate_motion_at_points(self, points) -> np.array:
+    def evaluate_motion_at_points(self, points: np.ndarray) -> np.ndarray:
         return np.cross(self.direction, points - self.rotation_center)
 
-    @lru_cache
-    def evaluate_gradient_of_motion(self, mesh) -> np.array:
-        grad = np.cross(self.direction, np.eye(3))
-        return np.tile(grad, (mesh.nb_faces, 1, 1))
+    def evaluate_gradient_of_motion_at_points(self, points: np.ndarray) -> np.ndarray:
+        grad = np.cross(self.direction, np.eye(3)).T
+        # Transposing because np.cross compute the cross product row-wise,
+        # but we want it column-wise for the conventions of the Jacobian matrix.
+        return np.tile(grad, (points.shape[0], 1, 1))
 
 
 class DofOnSubmesh(AbstractDof):
+    """Defines a dof that is zeros everywhere except on a given set of faces
+    on which another dof object is used for the definition.
+
+    Parameters
+    ----------
+    dof: AbstractDof
+        Some other dof
+    faces: boolean array (or slice or range)
+        The indices of the faces on which the `dof` should be evaluated.
+        Can be provided as a list of indices or a slice,
+        or as a boolean array of size `nb_faces` the total size of the mesh on which this dof is defined.
+    """
     def __init__(self, dof: AbstractDof, faces):
         self.dof = dof
         self.faces = faces
 
-    def evaluate_motion(self, mesh):
-        motion = np.zeros((mesh.nb_faces, 3))
-        motion[self.faces, :] = self.dof.evaluate_motion(mesh.extract_faces(self.faces))
+    def __str__(self):
+        return f"DofOnSubmesh(dof={self.dof}, faces={self.faces})"
+
+    def evaluate_motion_at_points(self, points: np.ndarray) -> np.ndarray:
+        motion = np.zeros((points.shape[0], 3))
+        motion[self.faces, :] = self.dof.evaluate_motion_at_points(points[self.faces, :])
         return motion
 
-    def evaluate_motion_at_points(self, points):
-        raise NotImplementedError()
-
-    def evaluate_gradient_of_motion(self, mesh) -> np.array:
-        grad = np.zeros((mesh.nb_faces, 3, 3))
-        grad[self.faces, :, :] = self.dof.evaluate_gradient_of_motion(mesh.extract_faces(self.faces))
+    def evaluate_gradient_of_motion_at_points(self, points: np.ndarray) -> np.ndarray:
+        grad = np.zeros((points.shape[0], 3, 3))
+        grad[self.faces, :, :] = self.dof.evaluate_gradient_of_motion_at_points(points[self.faces, :])
         return grad