+++ src/capytaine/meshes/abstract_meshes.py
@@ -171,14 +171,14 @@ class AbstractMesh(SurfaceIntegralsMixin, ABC):
 
         R = np.eye(3) + K + K @ K * ((1 - c) / (s ** 2))
         return self.rotated_with_matrix(R, name=name)
-    
-    def edges_faces_waterline(self): 
+
+    def edges_faces_waterline(self):
         """Extract the water line of the mesh.
 
         Returns
         -------
         (np.ndarray, np.ndarray)
-            A tuple (edges, faces) where the edges are the edges of the water line (pairs of vertex indices) 
+            A tuple (edges, faces) where the edges are the edges of the water line (pairs of vertex indices)
             and the faces are the corresponding faces of the mesh that contain these edges (quadruplet of vertex indices).
         """
         epsilon = 1e-6
@@ -193,22 +193,22 @@ class AbstractMesh(SurfaceIntegralsMixin, ABC):
                 faces_waterline.append(self.faces[k,:])
 
         return np.array(edges_waterline), np.array(faces_waterline)
-    
+
     @property
     def nb_edges_waterline(self) -> int:
         """Number of edges in the water line."""
         return np.shape(self.edges_waterline)[0]
-    
+
     @cached_property
     def edges_waterline(self):
         """Return an array with the edges of the water line as pairs of vertex indices with shape (nb_edges_waterline,2)."""
         return self.edges_faces_waterline()[0]
-    
+
     @cached_property
     def faces_waterline(self):
         """Return an array with the faces of the water line as quadruplets of vertex indices with shape (nb_edges_waterline,4)."""
         return self.edges_faces_waterline()[1]
-    
+
     @cached_property
     def length_edges_waterline(self):
         """Return an array with the lengths of the edges of the water line with shape (nb_edges_waterline,)."""
@@ -217,7 +217,7 @@ class AbstractMesh(SurfaceIntegralsMixin, ABC):
         vertices_right = self.vertices[edges[:,1],:]
         length_waterline = np.linalg.norm(vertices_left - vertices_right, ord=2, axis=1)
         return length_waterline
-    
+
     def waterline_integral(self, data):
         """Returns integral of given data along the water line.
 
@@ -229,7 +229,7 @@ class AbstractMesh(SurfaceIntegralsMixin, ABC):
         Returns
         -------
         float
-            Value of the integral. 
+            Value of the integral.
         """
         return np.sum(self.length_edges_waterline*data)